[Java] Split apart schema generation files.
I think separate files improves the readability of the arbitrary message
schema generation.
