+++ sbe-tool/src/propertyTest/java/uk/co/real_logic/sbe/properties/schema/TypeSchemaVisitor.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2013-2023 Real Logic Limited.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.co.real_logic.sbe.properties.schema;
+
+public interface TypeSchemaVisitor
+{
+    void onEncoded(EncodedDataTypeSchema type);
+
+    void onComposite(CompositeTypeSchema type);
+
+    void onEnum(EnumTypeSchema type);
+
+    void onSet(SetSchema type);
+}
